+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [8.2.4] - 2026-9-9
+
+### Updated og.js
+
 ## [8.2.3] - 2026-9-9
 
 ### Fixed Og Images for folders and courses